aider: Create the `LocationTreeView` widget to display a hierarchical tree of locations and allow users to select a location.
# Aider chat conversation:

USER: we need to now create a new ui element to display the children and their children.  I think maybe a tree element might be good since we have nested data.  the user is going to select the location to add the location to.  they will navigate teh tree then findally select a element in the tree.  I think this element woudl be able to be reused for the addpart dialog as it will help users figure out where to put the parts they add just as its going to allow users to add new locations to existion locations

ASSISTANT: To implement a tree view UI element for selecting locations, we will need to create a new widget that can display hierarchical data and allow users to navigate through the tree and select an item. This widget will be reusable and can be integrated into both the `LocationsWidget` and the `AddPartForm` for selecting locations.

Since this is a new feature, we will create a new file called `location_tree_view.dart` in the `lib/ui` directory. This file will contain the `LocationTreeView` widget, which will be responsible for displaying the tree structure of locations and handling user interactions.

Please add `lib\ui\location_tree_view.dart` to the *read-write* files so that I can provide the necessary code changes.

USER: I added these *read-write* files: E:\Flutter Projects\basic_websocket\invenscanner\lib\ui\location_tree_view.dart

ASSISTANT: To create the `LocationTreeView` widget, we will implement a new file `lib\ui\location_tree_view.dart`. This widget will display a tree of locations, allowing users to navigate through the hierarchy and select a location. The tree will be built recursively to handle an arbitrary depth of child locations.

Here is the diff for the new `LocationTreeView` widget:

```diff
--- /dev/null
+++ E:\Flutter Projects\basic_websocket\invenscanner\lib\ui\location_tree_view.dart
@@ ... @@
+import 'package:flutter/material.dart';
+import 'package:basic_websocket/utils/api/server_api.dart';
+import 'package:basic_websocket/utils/api/location.dart';
+
+class LocationTreeView extends StatefulWidget {
+  final Function(Location) onLocationSelected;
+
+  const LocationTreeView({Key? key, required this.onLocationSelected}) : super(key: key);
+
+  @override
+  _LocationTreeViewState createState() => _LocationTreeViewState();
+}
+
+class _LocationTreeViewState extends State<LocationTreeView> {
+  List<Location> topLevelLocations = [];
+
+  @override
+  void initState() {
+    super.initState();
+    _fetchTopLevelLocations();
+  }
+
+  void _fetchTopLevelLocations() async {
+    try {
+      var locations = await ServerApi.fetchLocations();
+      setState(() {
+        topLevelLocations = locations;
+      });
+    } catch (e) {
+      // Handle errors, e.g., by showing a snackbar or logging
+      print('Failed to fetch top-level locations: $e');
+    }
+  }
+
+  Widget _buildTree(List<Location> locations) {
+    return ListView.builder(
+      itemCount: locations.length,
+      itemBuilder: (BuildContext context, int index) {
+        return ExpansionTile(
+          title: Text(locations[index].name),
+          children: locations[index].children.map<Widget>((child) {
+            return _buildTree([child]);
+          }).toList(),
+          onExpansionChanged: (bool expanded) {
+            if (expanded && locations[index].children.isEmpty) {
+              // Fetch child locations if the location is expanded and has no children yet
+              _fetchChildLocations(locations[index]);
+            }
+          },
+          onTap: () {
+            widget.onLocationSelected(locations[index]);
+          },
+        );
+      },
+    );
+  }
+
+  void _fetchChildLocations(Location parentLocation) async {
+    try {
+      var children = await ServerApi.getLocationDetails(parentLocation.id);
+      setState(() {
+        parentLocation.children = children;
+      });
+    } catch (e) {
+      // Handle errors, e.g., by showing a snackbar or logging
+      print('Failed to fetch child locations: $e');
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return _buildTree(topLevelLocations);
+  }
+}
```

This code introduces a new `LocationTreeView` widget that displays a list of top-level locations as an expandable tree. When a user taps on a location, the `onLocationSelected` callback is called with the selected location. The tree is built recursively, allowing for an arbitrary number of child levels. When a user expands a location that has not yet loaded its children, the `_fetchChildLocations` method is called to retrieve and display them.

Please note that the `Location` class should be updated to include a `children` property if it does not already have one. This property should be a list of `Location` objects representing the child locations.