aider: Fix setting parent ID when adding new location with a parent location set.
# Aider chat conversation:

USER: it looks like in the locations.dart file that when adding a new location with a parent location set, its not adding the id of that location when using the `ServerApi` to create the new location

ASSISTANT: To fix this issue, we need to modify the `_addLocation` function in the `lib\ui\locations.dart` file. Currently, the parent ID is being set to the ID of the selected top-level location. However, if a child location is selected, its ID should be used as the parent ID for the new location.

Here is the *SEARCH/REPLACE* block:

lib\ui\locations.dart
```python
<<<<<<< SEARCH
    String? parentId;
    if (selectedTopLevelLocation?.id == "root") {
      parentId = null;
    } else {
      parentId = selectedTopLevelLocation?.id;
    }
=======
    String? parentId = selectedChildLocation?.id ?? selectedTopLevelLocation?.id;
    if (parentId == "root") {
      parentId = null;
    }
>>>>>>> REPLACE
```