aider: The null safety checks have been added to prevent potential null pointer exceptions.
# Aider chat conversation:

USER: that didnt work

ASSISTANT: I apologize for the confusion. Let's try to address the issue.

The error message you're seeing is due to the fact that the `currentState` property is being accessed on a `GlobalKey` that could potentially be `null`. To fix this, we need to add null safety checks before accessing the `currentState` property.

Here are the changes we need to make:

1. In the `lib\ui\locations.dart` file, we need to add a null safety check before accessing the `currentState` property in the `_addLocation` function.

2. In the `lib\ui\location_tree_view.dart` file, we need to add a null safety check before accessing the `currentState` property in the `refreshTree` function.

Here are the *SEARCH/REPLACE* blocks:

lib\ui\locations.dart
```dart
<<<<<<< SEARCH
      locationTreeView.key.currentState?.refreshTree();
=======
      if (locationTreeView.key.currentState != null) {
        locationTreeView.key.currentState!.refreshTree();
      }
>>>>>>> REPLACE
```

lib\ui\location_tree_view.dart
```dart
<<<<<<< SEARCH
    treeController = TreeController<LocationNode>(
=======
    if (treeController != null) {
      treeController = TreeController<LocationNode>(
    }
>>>>>>> REPLACE
```

Please note that these changes will only prevent the error from occurring. If the `currentState` is `null` at the time these lines of code are executed, the `refreshTree` function will not be called. You may need to ensure that the `currentState` is not `null` at the time these lines of code are executed.